#1334 move checkbox logic to relavant wrapper
