+++ lib/elementWrapper/checkBoxWrapper.js
@@ -1,7 +1,61 @@
 const CheckBox = require('../elements/checkBox');
+const { $function } = require('../elementSearch');
 const ElementWrapper = require('./elementWrapper');
-const { firstElement } = require('./helper');
+const { firstElement, getElementGetter } = require('./helper');
+
+function getCheckBoxElementWithLabel(label) {
+  const checkBoxes = [];
+  function checkAndPushElement(elem) {
+    if (
+      elem.tagName &&
+      elem.tagName.toLowerCase() == 'input' &&
+      elem.type &&
+      elem.type.toLowerCase() === 'checkbox'
+    ) {
+      checkBoxes.push(elem);
+    }
+  }
+  const matchingLabels = [...document.querySelectorAll('label')].filter((labelElem) => {
+    return labelElem.innerText.toLowerCase().includes(label.toLowerCase());
+  });
+  for (let matchingLabel of matchingLabels) {
+    const labelFor = matchingLabel.getAttribute('for');
+    if (labelFor) {
+      //check label with attribute for
+      const labelForElement = document.getElementById(labelFor);
+      checkAndPushElement(labelForElement);
+    } else {
+      // check child node of label tag
+      matchingLabel.childNodes.forEach((elem) => {
+        checkAndPushElement(elem);
+      });
+    }
+  }
+  //check label is inlined
+  const inputRadioElements = [...document.querySelectorAll('input[type="checkbox" i]')];
+  for (let inputRadioElement of inputRadioElements) {
+    if (
+      inputRadioElement.nextSibling.nodeType === Node.TEXT_NODE &&
+      inputRadioElement.nextSibling.wholeText.includes(label)
+    ) {
+      checkBoxes.push(inputRadioElement);
+    }
+  }
+  return checkBoxes;
+}
 class CheckBoxWrapper extends ElementWrapper {
+  _get = getElementGetter(
+    this.selector,
+    async () =>
+      await $function(
+        getCheckBoxElementWithLabel,
+        this.selector.label,
+        this.options.selectHiddenElements,
+      ),
+    'input[type="checkbox" i]',
+    this.options.selectHiddenElements,
+  );
+
   async isChecked() {
     const elem = await firstElement.apply(this);
     return await elem.isChecked();
