+++ lib/elementWrapper/elementWrapper.js
@@ -1,12 +1,12 @@
 const { firstElement } = require('./helper');
 
 const { descEvent } = require('../eventBus');
-let _getIfExists;
+let {getIfExists} = require('../elementSearch');
 class ElementWrapper {
-  constructor(get, description, getIfExists) {
-    this._get = get;
+  constructor(selector, options, description) {
+    this.selector = selector;
+    this.options = options;
     this._description = description;
-    _getIfExists = getIfExists;
   }
 
   async get(retryInterval, retryTimeout) {
@@ -50,11 +50,11 @@ class ElementWrapper {
     return await elem.isDisabled();
   }
   async elements(retryInterval, retryTimeout) {
-    return await _getIfExists(this._get, this._description)(null, retryInterval, retryTimeout);
+    return await getIfExists(this._get, this._description)(null, retryInterval, retryTimeout);
   }
 
   async element(index, retryInterval, retryTimeout) {
-    const results = await _getIfExists(this._get, this._description)(
+    const results = await getIfExists(this._get, this._description)(
       null,
       retryInterval,
       retryTimeout,
