+++ lib/elementWrapper/elementWrapperFactory.js
@@ -7,110 +7,33 @@ const DropDownWrapper = require('./dropDownWrapper');
 const ColorWrapper = require('./colorWrapper');
 const RangeWrapper = require('./rangeWrapper');
 const TimeFieldWrapper = require('./timeFieldWrapper');
+const { prepareParameters, desc } = require('./helper');
 
-const { xpath, isString, isElement, isSelector } = require('../helper');
-const { RelativeSearchElement, handleRelativeSearch } = require('../proximityElementSearch');
-const { $$, getIfExists } = require('../elementSearch');
-
-function hasProximitySelectors(values) {
-  return values.selector.args.length > 0;
-}
-
-const prepareParameters = (attrValuePairs, options, ...args) => {
-  if (options instanceof RelativeSearchElement) {
-    args = [options].concat(args);
-    options = {};
-  }
-  let values = {
-    selector: { attrValuePairs: attrValuePairs, args: args },
-  };
-  if (attrValuePairs instanceof RelativeSearchElement) {
-    args = [attrValuePairs].concat(args);
-    values.selector = { args: args };
-  } else if (isString(attrValuePairs) || isSelector(attrValuePairs) || isElement(attrValuePairs)) {
-    values.selector = { label: attrValuePairs, args: args };
-  }
-  values.options = options || {};
-  if (values.options.selectHiddenElements && hasProximitySelectors(values)) {
-    console.warn(
-      'WARNING: Proximity of hidden element are not available hence proximity selector will be ignored',
-    );
-  }
-  return values;
-};
-
-const getQuery = (attrValuePairs, tag = '') => {
-  let path = tag;
-  for (const key in attrValuePairs) {
-    if (key === 'class') {
-      path += `[${key}~="${attrValuePairs[key]}"]`;
-    } else {
-      path += `[${key}="${attrValuePairs[key]}"]`;
-    }
-  }
-  return path;
-};
-
-const getElementGetter = (selector, query, tags, selectHiddenElements) => {
-  let get;
-  if (selector.attrValuePairs) {
-    let query = tags
-      .split(',')
-      .map((tag) => getQuery(selector.attrValuePairs, tag))
-      .join(',');
-    get = async () =>
-      await handleRelativeSearch(await $$(query, selectHiddenElements), selector.args);
-  } else if (selector.label) {
-    get = async () => await handleRelativeSearch(await query(), selector.args);
-  } else {
-    get = async () =>
-      await handleRelativeSearch(await $$(tags, selectHiddenElements), selector.args);
-  }
-  return get;
-};
-
-const desc = (selector, query, tag) => {
-  let description = '';
-  if (selector.attrValuePairs) {
-    description = getQuery(selector.attrValuePairs, tag);
-  } else if (selector.label) {
-    description = `${tag} with ${query} ${selector.label} `;
-  }
-
-  for (const arg of selector.args) {
-    description += description === '' ? tag : ' and';
-    description += ' ' + arg.toString();
-  }
-
-  return description;
-};
-
-const createElementWrapper = (get, description, elementType) => {
+const createElementWrapper = (elementType, attrValuePairs, _options, ...args) => {
+  const { selector, options } = prepareParameters(attrValuePairs, _options, ...args);
+  const description = desc(selector, 'label', 'Checkbox');
   switch (elementType) {
-    case 'fileField':
-      return new FileFieldWrapper(get, description, getIfExists);
-    case 'textBox':
-      return new TextBoxWrapper(get, description, getIfExists);
-    case 'radioButton':
-      return new RadioButtonWrapper(get, description, getIfExists);
+    // case 'fileField':
+    //   return new FileFieldWrapper(get, description, getIfExists);
+    // case 'textBox':
+    //   return new TextBoxWrapper(get, description, getIfExists);
+    // case 'radioButton':
+    //   return new RadioButtonWrapper(get, description, getIfExists);
     case 'checkBox':
-      return new CheckBoxWrapper(get, description, getIfExists);
-    case 'dropDown':
-      return new DropDownWrapper(get, description, getIfExists);
-    case 'colorField':
-      return new ColorWrapper(get, description, getIfExists);
-    case 'rangeType':
-      return new RangeWrapper(get, description, getIfExists);
-    case 'timeField':
-      return new TimeFieldWrapper(get, description, getIfExists);
-    default:
-      return new ElementWrapper(get, description, getIfExists);
+      return new CheckBoxWrapper(selector, options, description);
+    // case 'dropDown':
+    //   return new DropDownWrapper(get, description, getIfExists);
+    // case 'colorField':
+    //   return new ColorWrapper(get, description, getIfExists);
+    // case 'rangeType':
+    //   return new RangeWrapper(get, description, getIfExists);
+    // case 'timeField':
+    //   return new TimeFieldWrapper(get, description, getIfExists);
+    // default:
+    //   return new ElementWrapper(get, description, getIfExists);
   }
 };
 
 module.exports = {
-  prepareParameters,
-  getElementGetter,
-  desc,
   createElementWrapper,
 };
