+++ lib/helper.js
@@ -23,8 +23,6 @@ const path = require('path');
 let projectRoot = null;
 const timeouts = [];
 const descEvent = require('./eventBus').descEvent;
-const Element = require('./elements/element');
-const ElementWrapper = require('./elementWrapper/elementWrapper');
 
 class Helper {
   /**
@@ -164,12 +162,14 @@ const handleUrlRedirection = (url) => {
 };
 
 const isSelector = (obj) =>
-  (obj &&
-    Object.prototype.hasOwnProperty.call(obj, 'elements') &&
-    Object.prototype.hasOwnProperty.call(obj, 'exists')) ||
-  obj instanceof ElementWrapper;
+  obj &&
+  Object.prototype.hasOwnProperty.call(obj, 'elements') &&
+  Object.prototype.hasOwnProperty.call(obj, 'exists');
 
-const isElement = (obj) => obj && obj instanceof Element;
+const isElement = (obj) =>
+  obj &&
+  Object.prototype.hasOwnProperty.call(obj, 'get') &&
+  Object.prototype.hasOwnProperty.call(obj, 'text');
 
 module.exports = {
   helper: Helper,
