+++ lib/taiko.js
@@ -32,7 +32,6 @@ const {
   $$xpath,
   findElements,
   findFirstElement,
-  typeEqual,
   $function,
 } = require('./elementSearch');
 const { handleRelativeSearch, RelativeSearchElement } = require('./proximityElementSearch');
@@ -40,12 +39,8 @@ const Element = require('./elements/element');
 const { BrowserContext } = require('./browserContext');
 const keyDefinitions = require('../lib/data/USKeyboardLayout');
 
-const {
-  prepareParameters,
-  getElementGetter,
-  desc,
-  createElementWrapper,
-} = require('./elementWrapper/elementWrapperFactory');
+const { prepareParameters, desc, getElementGetter } = require('./elementWrapper/helper');
+const { createElementWrapper } = require('./elementWrapper/elementWrapperFactory');
 const {
   setConfig,
   getConfig,
@@ -2814,57 +2809,7 @@ function dropDown(attrValuePairs, _options, ...args) {
  */
 module.exports.checkBox = (attrValuePairs, _options, ...args) => {
   validate();
-  const { selector, options } = prepareParameters(attrValuePairs, _options, ...args);
-  function getCheckBoxElementWithLabel(label) {
-    const checkBoxes = [];
-    function checkAndPushElement(elem) {
-      if (
-        elem.tagName &&
-        elem.tagName.toLowerCase() == 'input' &&
-        elem.type &&
-        elem.type.toLowerCase() === 'checkbox'
-      ) {
-        checkBoxes.push(elem);
-      }
-    }
-    const matchingLabels = [...document.querySelectorAll('label')].filter((labelElem) => {
-      return labelElem.innerText.toLowerCase().includes(label.toLowerCase());
-    });
-    for (let matchingLabel of matchingLabels) {
-      const labelFor = matchingLabel.getAttribute('for');
-      if (labelFor) {
-        //check label with attribute for
-        const labelForElement = document.getElementById(labelFor);
-        checkAndPushElement(labelForElement);
-      } else {
-        // check child node of label tag
-        matchingLabel.childNodes.forEach((elem) => {
-          checkAndPushElement(elem);
-        });
-      }
-    }
-    //check label is inlined
-    const inputRadioElements = [...document.querySelectorAll('input[type="checkbox" i]')];
-    for (let inputRadioElement of inputRadioElements) {
-      if (
-        inputRadioElement.nextSibling.nodeType === Node.TEXT_NODE &&
-        inputRadioElement.nextSibling.wholeText.includes(label)
-      ) {
-        checkBoxes.push(inputRadioElement);
-      }
-    }
-    return checkBoxes;
-  }
-  const get = getElementGetter(
-    selector,
-    async () =>
-      await $function(getCheckBoxElementWithLabel, selector.label, options.selectHiddenElements),
-    'input[type="checkbox" i]',
-    options.selectHiddenElements,
-  );
-
-  const description = desc(selector, 'label', 'Checkbox');
-  return createElementWrapper(get, description, 'checkBox');
+  return createElementWrapper('checkBox', attrValuePairs, _options, ...args);
 };
 
 /**
